gindex.php
new index page for basic transaction. not connecting to the database yet. I think I might need to make the table first.

And figure out what the hell <form action="index.php" method="post"> fucking means, is that the file that the form is on? is it the file the info is sent to? god damn it why is this difficult for me. post. posting, sending maybe retrieving? fuck...