Create WORKSONMSI_index.php
This works on my MSI in that I can add to my Database on the table that i wanted. There is still an error in the PHPstorm for the $con variable undefined, but it still works to add things to that table from the web interface of localhost, might change the ports.